requirements.txt
absl-py==2.0.0
addict==2.4.0
aiohttp==3.8.6
aiosignal==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
altgraph==0.17.4
aniso8601==9.0.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
bidict==0.22.1
boltons @ file:///C:/b/abs_707eo7c09t/croot/boltons_1677628723117/work
Bottleneck==1.3.7
brotlipy==0.7.0
cachetools==5.3.1
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
coloredlogs==14.0
conda==23.3.1
conda-content-trust @ file:///C:/Windows/TEMP/abs_4589313d-fc62-4ccc-81c0-b801b4449e833j1ajrwu/croots/recipe/conda-content-trust_1658126379362/work
conda-package-handling @ file:///C:/b/abs_fcga8w0uem/croot/conda-package-handling_1672865024290/work
conda_package_streaming @ file:///C:/b/abs_0e5n5hdal3/croot/conda-package-streaming_1670508162902/work
crcmod==1.7
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
Cython==0.29.36
datasets==2.13.0
decorator==5.1.1
dill==0.3.6
Distance==0.1.3
dnspython==2.3.0
edit-distance==1.0.6
editdistance==0.6.2
einops==0.7.0
espnet-tts-frontend==0.0.3
et-xmlfile==1.1.0
eventlet==0.33.3
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpeg==1.4
filelock==3.12.4
Flask==2.2.5
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
flask-talisman==1.1.0
flatbuffers==24.3.25
frozenlist==1.4.0
fsspec==2023.9.2
funasr==1.1.16
g2p==1.1.20230822
g2p-en==2.1.0
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
greenlet==3.0.0
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
hdbscan==0.8.33
huggingface-hub==0.17.3
humanfriendly==10.0
hydra-core==1.3.2
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
importlib-metadata==6.8.0
inflect==7.0.0
itsdangerous==2.1.2
jaconv==0.3.4
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
kaldiio==2.18.0
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.0
loguru==0.7.3
Markdown==3.5
MarkupSafe==2.1.3
menuinst @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_455sf5o0ct/croots/recipe/menuinst_1661805970842/work
modelscope==1.21.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.14
munkres==1.1.4
nara-wpe==0.0.9
networkx==2.8.4
nltk==3.8.1
numba==0.58.0
numpy==1.25.2
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.20.1
openpyxl==3.1.2
oss2==2.18.2
packaging @ file:///C:/b/abs_ed_kb9w6g4/croot/packaging_1678965418855/work
pandas==2.1.1
panphon==0.20.0
pefile==2023.2.7
Pillow==10.1.0
platformdirs==3.11.0
pluggy @ file:///C:/ci/pluggy_1648042746254/work
pooch==1.7.0
protobuf==4.24.4
psutil==7.0.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycosat @ file:///C:/b/abs_4b1rrw8pn9/croot/pycosat_1666807711599/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pynndescent==0.5.11
pyOpenSSL @ file:///C:/b/abs_552w85x1jz/croot/pyopenssl_1677607703691/work
pypinyin==0.49.0
pyreadline3==3.4.1
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
pystray==0.19.5
python-dateutil==2.8.2
python-engineio==4.8.0
python-socketio==5.10.0
pytorch-wpe==0.0.1
pytz==2023.3.post1
pywin32-ctypes==0.2.3
PyYAML==6.0.1
regex==2023.10.3
requests @ file:///C:/b/abs_80_edxtvhu/croot/requests_1678709752318/work
requests-oauthlib==1.3.1
rotary-embedding-torch==0.3.2
rsa==4.9
ruamel.yaml @ file:///C:/b/abs_30ee5qbthd/croot/ruamel.yaml_1666304562000/work
ruamel.yaml.clib @ file:///C:/b/abs_aarblxbilo/croot/ruamel.yaml.clib_1666302270884/work
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
sentencepiece==0.1.99
simple-websocket==1.0.0
simplejson==3.19.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
starlette==0.41.3
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
text-unidecode==1.3
TextGrid==1.5
threadpoolctl==3.2.0
tokenizers==0.14.1
tomli==2.0.1
toolz @ file:///C:/b/abs_cfvk6rc40d/croot/toolz_1667464080130/work
torch==2.1.0+cu121
torch-complex==0.4.3
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121
tqdm @ file:///C:/b/abs_f76j9hg7pv/croot/tqdm_1679561871187/work
transformers==4.34.0
typing_extensions==4.8.0
tzdata==2023.3
umap==0.1.1
umap-learn==0.5.5
unicodecsv==0.14.1
Unidecode==1.3.7
urllib3 @ file:///C:/b/abs_3ce53vrdcr/croot/urllib3_1680254693505/work
uvicorn==0.34.0
websockets==12.0
Werkzeug==2.2.3
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
win32_setctime==1.2.0
wincertstore==0.2
wsproto==1.2.0
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0
zstandard==0.19.0
